@@ -1,7 +1,7 @@
 package com.mcmiddleearth.mcmetours.proxy.bungee.listener;
 
-import com.mcmiddleearth.mcmetours.proxy.core.util.PluginData;
-import com.mcmiddleearth.mcmetours.proxy.core.tour.Tour;
+import com.mcmiddleearth.base.bungee.player.BungeeMcmePlayer;
+import com.mcmiddleearth.mcmetours.proxy.core.eventHandler.ChatHandler;
 import net.md_5.bungee.api.connection.ProxiedPlayer;
 import net.md_5.bungee.api.event.ChatEvent;
 import net.md_5.bungee.api.plugin.Listener;
@@ -9,22 +9,14 @@ import net.md_5.bungee.event.EventHandler;
 import net.md_5.bungee.event.EventPriority;
 
 /**
- * @author Jubo
+ * @author Jubo, Eriol_Eandur
  */
 public class ChatListener implements Listener {
 
     @EventHandler(priority = EventPriority.NORMAL)
     public void TourChat(ChatEvent event){
-        ProxiedPlayer sender = (ProxiedPlayer) event.getSender();
-        String message = event.getMessage();
-        if(!message.startsWith("/")){
-            Tour tour = PluginData.getTour(sender);
-            if(tour != null ) {
-                if(tour.getTourChat().contains(sender)) {
-                    tour.tourChat(sender,message);
-                    event.setCancelled(true);
-                }
-            }
+        if(ChatHandler.handle(new BungeeMcmePlayer((ProxiedPlayer) event.getSender()), event.getMessage())) {
+            event.setCancelled(true);
         }
     }
 }
