@@ -1,27 +1,27 @@
 package com.mcmiddleearth.mcmetours.proxy.core.command.arguments;
 
-import com.mcmiddleearth.command.argument.AbstractPlayerArgumentType;
-import com.mcmiddleearth.mcmetours.proxy.core.command.TourCommandSender;
+import com.mcmiddleearth.base.core.command.argument.AbstractPlayerArgumentType;
+import com.mcmiddleearth.base.core.player.McmeProxyPlayer;
+import com.mcmiddleearth.mcmetours.proxy.core.McmeTours;
 import com.mcmiddleearth.mcmetours.proxy.core.util.Permission;
 import com.mcmiddleearth.mcmetours.proxy.core.util.PluginData;
 import com.mojang.brigadier.LiteralMessage;
 import com.mojang.brigadier.StringReader;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import com.mojang.brigadier.exceptions.SimpleCommandExceptionType;
-import net.md_5.bungee.api.ProxyServer;
 
 import java.util.Collection;
 import java.util.HashSet;
 
 /**
- * @author Jubo
+ * @author Jubo, Eriol_Eandur
  */
 public class CommandPermissionArgument extends AbstractPlayerArgumentType {
 
     @Override
     public String parse(StringReader reader) throws CommandSyntaxException{
         String o = reader.readUnquotedString();
-        TourCommandSender sender = new TourCommandSender(ProxyServer.getInstance().getPlayer(o));
+        McmeProxyPlayer sender = McmeTours.getProxy().getPlayer(o);
         if(PluginData.getTour(sender) != null && PluginData.hasPermission(sender, Permission.HOST)){
             return o;
         }
