@@ -1,7 +1,8 @@
 package com.mcmiddleearth.mcmetours.proxy.core.command.arguments;
 
-import com.mcmiddleearth.command.argument.AbstractPlayerArgumentType;
-import com.mcmiddleearth.mcmetours.proxy.core.command.TourCommandSender;
+import com.mcmiddleearth.base.core.command.argument.AbstractPlayerArgumentType;
+import com.mcmiddleearth.base.core.player.McmeProxyPlayer;
+import com.mcmiddleearth.mcmetours.proxy.core.McmeTours;
 import com.mcmiddleearth.mcmetours.proxy.core.util.PluginData;
 import com.mojang.brigadier.LiteralMessage;
 import com.mojang.brigadier.StringReader;
@@ -12,14 +13,14 @@ import java.util.Collection;
 import java.util.stream.Collectors;
 
 /**
- * @author Jubo
+ * @author Jubo, Eriol_Eandur
  */
 public class CommandPlayerArgument extends AbstractPlayerArgumentType {
 
     @Override
     public String parse(StringReader reader) throws CommandSyntaxException {
         String o = reader.readUnquotedString();
-        TourCommandSender sender = new TourCommandSender(ProxyServer.getInstance().getPlayer(o));
+        McmeProxyPlayer sender = McmeTours.getProxy().getPlayer(o);
         if(PluginData.getTour(sender) != null){
             return o;
         }
@@ -29,6 +30,6 @@ public class CommandPlayerArgument extends AbstractPlayerArgumentType {
 
     @Override
     protected Collection<String> getPlayerSuggestions() {
-        return ProxyServer.getInstance().getPlayers().stream().map(ProxiedPlayer::getName).collect(Collectors.toSet());
+        return McmeTours.getProxy().getPlayers().stream().map(McmeProxyPlayer::getName).collect(Collectors.toSet());
     }
 }
