@@ -1,10 +1,11 @@
 package com.mcmiddleearth.mcmetours.proxy.core.command.handler;
 
-import com.mcmiddleearth.command.AbstractCommandHandler;
-import com.mcmiddleearth.command.McmeCommandSender;
-import com.mcmiddleearth.command.builder.HelpfulLiteralBuilder;
-import com.mcmiddleearth.command.builder.HelpfulRequiredArgumentBuilder;
-import com.mcmiddleearth.mcmetours.proxy.core.command.TourCommandSender;
+import com.mcmiddleearth.base.core.command.handler.AbstractCommandHandler;
+import com.mcmiddleearth.base.core.command.McmeCommandSender;
+import com.mcmiddleearth.base.core.command.builder.HelpfulLiteralBuilder;
+import com.mcmiddleearth.base.core.command.builder.HelpfulRequiredArgumentBuilder;
+import com.mcmiddleearth.base.core.player.McmeProxyPlayer;
+import com.mcmiddleearth.base.core.plugin.McmePlugin;
 import com.mcmiddleearth.mcmetours.proxy.core.tour.Tour;
 import com.mcmiddleearth.mcmetours.proxy.core.util.Permission;
 import com.mcmiddleearth.mcmetours.proxy.core.util.PluginData;
@@ -12,25 +13,28 @@ import com.mcmiddleearth.mcmetours.proxy.core.util.PluginData;
 import static com.mojang.brigadier.arguments.StringArgumentType.greedyString;
 
 /**
- * @author Jubo
+ * @author Jubo, Eriol_Eandur
  */
 public class TcComandHandler extends AbstractCommandHandler {
 
-    public TcComandHandler(String name){super(name);}
+    public TcComandHandler(String name, McmePlugin plugin){super(name, plugin);}
 
     @Override
     protected HelpfulLiteralBuilder createCommandTree(HelpfulLiteralBuilder helpfulLiteralBuilder){
         helpfulLiteralBuilder
-                .requires(sender -> (PluginData.hasPermission((TourCommandSender) sender, Permission.USER) && PluginData.isInTour((TourCommandSender) sender)))
+                .requires(sender -> (sender instanceof McmeProxyPlayer player)
+                                        && (PluginData.hasPermission(sender, Permission.USER)
+                                        && PluginData.isInTour(player)))
                 .then(HelpfulRequiredArgumentBuilder.argument("text",greedyString())
                 .executes(context -> sendToTourChat(context.getSource(),context.getArgument("text",String.class))));
         return helpfulLiteralBuilder;
     }
 
     private int sendToTourChat(McmeCommandSender sender, String message){
-        Tour tour = PluginData.getTour((TourCommandSender) sender);
+        McmeProxyPlayer player = (McmeProxyPlayer) sender;
+        Tour tour = PluginData.getTour(player);
         if(tour != null ){
-            tour.tourChat((ProxiedPlayer) ((TourCommandSender) sender).getCommandSender(),message);
+            tour.tourChat(player, message);
         }
         return 0;
     }
