@@ -1,40 +1,34 @@
 package com.mcmiddleearth.mcmetours.proxy.core.tour;
 
 import com.mcmiddleearth.base.core.player.McmeProxyPlayer;
-import com.mcmiddleearth.command.McmeCommandSender;
 import com.mcmiddleearth.mcmetours.proxy.core.McmeTours;
-import com.mcmiddleearth.mcmetours.proxy.core.command.TourCommandSender;
+import com.mcmiddleearth.mcmetours.proxy.core.discord.TourDiscordHandler;
+import com.mcmiddleearth.mcmetours.proxy.core.util.MessageUtil;
 import com.mcmiddleearth.mcmetours.proxy.core.util.Permission;
 import com.mcmiddleearth.mcmetours.proxy.core.util.PluginData;
-import com.mcmiddleearth.mcmetours.proxy.core.discord.TourDiscordHandler;
 
 /**
- * @author Jubo
+ * @author Jubo, Eriol_Eandur
  */
 public class TourRequest {
 
-    public static void informHosts(McmeCommandSender sender, String message){
+    public static void informHosts(McmeProxyPlayer sender, String message){
         boolean success = false;
         if(PluginData.tourRunning()){
-            PluginData.getMessageUtil().sendClickableInfoMessage((McmeProxyPlayer)((TourCommandSender)sender).getCommandSender(), Style.INFO+PluginData.getMessageUtil().getPrefix()
-                    +"There is currently a tour running. "+Style.HIGHLIGHT+"Click here"+Style.INFO+" for more information","/tour check");
+            sender.sendMessage(MessageUtil.runningTourInfo());
             return;
         }
         for(McmeProxyPlayer player: McmeTours.getProxy().getPlayers()){
             if(player.hasPermission(Permission.HOST.getPermissionNode())){
-                if(message == null){
-                    PluginData.getMessageUtil().sendInfoMessage(player,TourCommandSender.getName((TourCommandSender) sender)+" has requested a tour!");
-                }else {
-                    PluginData.getMessageUtil().sendInfoMessage(player,TourCommandSender.getName((TourCommandSender) sender)+" has requested a tour! About: "+message);
-                }
+                sender.sendMessage(MessageUtil.tourRequest(player, message));
                 success = true;
             }
         }
         if(success){
-            PluginData.getMessageUtil().sendInfoMessage(sender, "A guide was informed about your request.");
+            sender.sendMessage(McmeTours.infoMessage("A guide was informed about your request."));
         }else{
-            PluginData.getMessageUtil().sendErrorMessage(sender,"There are currently no guides online.");
+            sender.sendMessage(McmeTours.infoMessage("There are currently no guides online."));
         }
-        new TourDiscordHandler((ProxiedPlayer) ((TourCommandSender)sender).getCommandSender(),null).requestTour(message);
+        new TourDiscordHandler(sender,null).requestTour(message);
     }
 }
